feat:大地粉砕1スキルを削除し、神力開放スキルにマージ
