--- Assets/Scripts/Data/Skill/GroundBrakerSkill_1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f931a9915ce1dd64ab97602606bd16c3
